requirements.txt
torch
torchvision
linear_attention_transformer
einops
typeguard
tensorboard
tqdm
numpy
pandas
scikit-learn
matplotlib
seaborn
lightning
torchinfo
pillow
mne[hdf5]
PyWavelets
PyYAML
joblib
scipy
x-transformers
pytest
rich
"git+https://github.com/bluorion-com/ZClip.git#egg=zclip[lightning]"